Add tests for missing logs & important systemd unit. Move test functions in util
